requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers==4.44.0
accelerate>=0.20.0
peft>=0.6.0
datasets>=2.14.0
pycocoevalcap
pycocotools
pandas
numpy
matplotlib
seaborn
Pillow
tqdm
scikit-learn
nltk
